Prevent HTTP request smuggling via Header normalization
As identified in RUSTSEC-2020-0031, normalizing the value of a header
field (through the use of `str::trim`) can make applications based on
this library vulnerable to HTTP request smuggling if the immediate
upstream load balancer interprets the malformed header in a different
way.

This backported patch based on a PR opened on the main tiny-http repo. [1]

[1]: https://github.com/tiny-http/tiny-http/pull/190